Update
Page Controller was updated and the patients home page data was changed so that today's date it automatically set but the patient can change it. Right now I'm trying to get an associative array to have the data work properly on the page and join. Another thing that it being worked on for this page is to have the patient's id and their name appear automatically on the page.